fix(create-knyt): embed package.json instead of reading it
Fixes an issue in the `create-knyt` CLI bundle by embedding the necessary properties from the package.json instead of resolving the module and reading it at runtime.

This issue was due to Bun's bundler resolving the module path given to `require.resolve` at build time, and embedding the result. The expected behavior was that the given module path would be resolved at runtime, not build time.