refactor(ssr): add assertion to `defineElementDefinition` to ensure `customElements` is available
This is effectively a refactor, because the function already assumes that `customElements` is available, but this makes it explicit and will throw a clear error if it's not. This is particularly useful for server-side rendering (SSR) scenarios where `customElements` might not be defined, if Knyt Glazier wasn't properly set up.