refactor: remove unnecessary dependencies (#17)
`@noble/hashes` was being used to generate hashes for comparing the contents of style sheets. The plan initially was to replace the package with native `SubtleCrypto` APIs, but on second thought, the hashing was unnecessary altogether. A simple string comparison is sufficient for our use case, so we can eliminate the dependency entirely. 😐 (doh)

The benefit of hashing the contents was to avoid comparing large strings directly, but in practice, it's negligible until the style sheets are megabytes in size, which is highly unlikely when these are typically only going to contain styles for a single component.

BREAKING CHANGE: Removed `hash` property. Style sheet contents are now compared directly as strings instead of using hashes.

Removed unnecessary dependencies for style sheet comparison.